feat: added DNI validation
