+++ src/pages/login.tsx
@@ -1,84 +1,13 @@
 import { NextPage } from "next";
 import Head from "next/head";
 import { useRouter } from "next/router";
-import { useEffect, useReducer, useState } from "react";
-import { twMerge } from "tailwind-merge";
+import { useEffect } from "react";
+import DniInput, { validateDNI } from "~/components/utils/form/DniInput";
+import { Input } from "~/components/utils/form/Input";
 import useAuth from "~/hooks/auth/useAuth";
 import useLogin, { Credentials } from "~/hooks/auth/useLogin";
+
 import useErrors from "~/hooks/utils/useErrors";
-
-type InputProps = React.DetailedHTMLProps<
-    React.InputHTMLAttributes<HTMLInputElement>,
-    HTMLInputElement
-> & {
-    error?: boolean;
-};
-
-const Input = ({ type, placeholder, error, onKeyDown }: InputProps) => {
-    const [showPassword, setShowPassword] = useState(false);
-
-    return (
-        <div
-            className={twMerge(
-                "input w-full flex items-center justify-end border-2 border-transparent",
-                error && "error",
-            )}
-        >
-            <input
-                className="relative w-full border-none bg-transparent outline-none"
-                type={
-                    type !== "password"
-                        ? type
-                        : showPassword
-                        ? "text"
-                        : "password"
-                }
-                placeholder={placeholder}
-                onKeyDown={onKeyDown}
-            />
-            {type === "password" && (
-                <span
-                    className="border-red border h-5 w-5 absolute cursor-pointer"
-                    onMouseDown={() => setShowPassword(true)}
-                    onMouseUp={() => setShowPassword(false)}
-                />
-            )}
-        </div>
-    );
-};
-
-const DniInput = ({ error }: { error: boolean }) => {
-    const [dni, setDni] = useState("");
-
-    const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-        if (e.key === "Backspace") {
-            if (dni.length === 0) return;
-            setDni(dni.slice(0, dni.length - 1));
-        } else if (e.key > "0" && e.key < "9") {
-            if (dni.length === 8) return;
-            setDni(dni + e.key);
-        } else if (e.key !== "Enter" && e.key !== "Tab") {
-            e.preventDefault();
-        }
-    };
-
-    return (
-        <div className="w-full flex items-center justify-start relative">
-            <Input
-                type="text"
-                placeholder={dni.length === 0 ? "DNI" : ""}
-                value={`${dni.slice(0, 2)}${
-                    dni.length > 2 ? "." : ""
-                }${dni.slice(2, 5)}${dni.length > 5 ? "." : ""}${dni.slice(
-                    5,
-                    8,
-                )}`}
-                onKeyDown={handleKeyDown}
-                error={error}
-            />
-        </div>
-    );
-};
 
 const Login: NextPage = () => {
     const router = useRouter();
@@ -117,9 +46,17 @@ const Login: NextPage = () => {
         e.preventDefault();
         setErrors({ type: "clear", all: true });
 
-        setTimeout(() => {
+        const dni = e.currentTarget.elements.namedItem(
+            "dni",
+        ) as HTMLInputElement;
+        const pass = e.currentTarget.elements.namedItem(
+            "pass",
+        ) as HTMLInputElement;
+
+        if (!validateDNI(dni.value)) {
             setErrors({ type: "set", input: "dni" });
-        }, 50);
+            return;
+        }
     };
 
     return (
@@ -141,11 +78,17 @@ const Login: NextPage = () => {
                             <h2 className="text-4xl font-semibold mb-5">
                                 Ingresar
                             </h2>
-                            <DniInput error={errors.dni} />
+                            <DniInput
+                                error={errors.dni}
+                                clearError={() =>
+                                    setErrors({ type: "clear", input: "dni" })
+                                }
+                            />
                             <Input
                                 type="password"
                                 placeholder="Contraseña"
                                 error={errors.pass}
+                                name="pass"
                             />
                             <p className="text-sm underline cursor-pointer self-center">
                                 No recuerdo mi contraseña
